Always parenthesize the type in a cast expression.
Consider this expression:

    4 as i32 < 42

The Rust parser is looking for a type following "as", and when it sees
the '<', it expects to find a list of type parameters. Then it reports a
parse error when it encounters "42", which is not a type.

Printing parentheses around the type name means it always knows where
the type ends and can return to looking for binary operators.

Found with csmith.