Lazily convert between bool and int.
My previous pattern for converting bool to int was:

    if b { 1 } else { 0 }

However, if that expression is used in a context where the result is not
needed, then Rust reports a type error: An if-expression as a statement
must have type (), but this expression has some unspecified integer type.

The key idea for fixing that bug is that when the result is not needed,
we don't need to convert from bool to int anyway, because the result
type is not observable if you don't use the result.

So this commit adds a CType variant to represent boolean subexpressions,
and only converts them back to int when the containing expression
requires that.

The next trick is that my previous invariant was that every expression
was translated to one of integer type, which meant I had to annotate any
operator which has a bool result in Rust using my `fromBool` helper.
Doing this lazily instead means that conversion is required for exactly
those operators that follow the "usual arithmetic conversions", so this
change simplifies some of the expression evaluation code.

Finally, it turns out that I didn't need that complicated if-expression,
because this cast is equivalent in Rust:

    b as i32

Which means adding a special case to the integer promotion rules
(notably, a case which is already required by the C standard)
automatically inserts the right bool-to-int conversion expression
everywhere it's needed.

Found with csmith.