Translate C types of integer literals.
In C, an expression like
    (unsigned) -2
is OK, because 2 is signed and negation makes sense on signed numbers.

If we translate this to Rust like so:
    -2 as u32
then Rust reports a "unary negation of unsigned integer" error, because
the literal's type wasn't specified so it infers the type of the cast,
and negation doesn't make sense on unsigned numbers.

Instead, we need to translate it as
    -2i32 as u32
which overrides the inferred type and then separately casts the value.

There might be other cases where C semantics require that the literal be
translated with an explicit Rust type or else the result would be wrong,
but I can't think of any. My implementation of the usual arithmetic
conversions should have taken care of it for binary operators, and I
don't think any other unary operators care about signedness.

Found with csmith.