Use wrapping operators for unsigned types.
In C, overflow is defined to wrap on unsigned types. In Rust, overflow
is a program error which panics in debug builds, but you can request
wrapping operations by using various "wrapping_*" method calls on the
various integer types. To reflect C's behavior, use the wrapping
variants for any arithmetic on unsigned types.

For now, I'm still translating signed arithmetic to the no-overflow
operators. Signed overflow has undefined behavior in C, and so I'm
within my rights to interpret it as something which should panic if it
happens at runtime. This is kind of like getting ubsan for free.

As future work, though, Brian Mastenbrook (@bmastenbrook) points out
that it would be useful to allow the user to request wrapping behavior
for signed overflow as well. Major C compilers have an option like
"-fwrapv" to request this behavior, and some projects rely on it.

Found with csmith, which generated unary negation operators applied to
unsigned expressions. Rust statically refuses to compile such an
expression. Thanks to https://twitter.com/spun_off/ for helping me
understand that this specific example was just an instance of C's
overflow rules, which I'd forgotten about.