Handle un-demanded ternary conditionals.
We translate ternary conditionals to Rust if/else expressions. If the
result of the ternary conditional is not used (meaning it's evaluated
only for its side effects), then Rust requires that the type of both
branches be `()`. That means we need to treat any final expression
within the block as a statement (i.e., put a semicolon after it).

Found with csmith, whose "safe_math_macros.h" can use ternary
conditionals in positions where the result of the safe-math operation is
unused. But I rather expect this to come up in some real code too.