Unary plus needs to pass `demand` through.
Unary plus doesn't translate to any code in Rust unless it forces an
integer promotion cast.

If its subexpression escapes unscathed, then when `demand` is False, the
subexpression becomes a statement in its own right.

If the subexpression is a ternary conditional, then this results in an
if/else statement which Rust, in this case, requires to have type ().
But if we passed True for `demand` when evaluating the subexpression,
then the if/else expression may have a type which is not ().

So when `demand` is already True, passing it through doesn't change the
behavior. When `demand` is False, given that I think we can get away
with skipping the integer promotion cast if the result isn't used,
pretending like the unary-plus operator just didn't exist gives us the
best result.

Found with csmith.