Readable error messages for all translation woes.
I previously used various mechanisms to catch unexpected or unsupported
constructs in the C source. All of them would eventually call `error` to
terminate the program, but they produced varying amounts of information
about what went wrong.

In cases where I explicitly called `error`, I usually included some
chunk of the C AST in the output, and that meant you could usually find
the file and line that triggered the error, and figure out which
construct was the problem. But I used the AST's `Show` instances, which
gave output that was very difficult to read.

In other cases, I used one of these mechanisms:

- lazy irrefutable pattern match
- anonymous lambda with a refutable pattern
- let-binding with a refutable pattern
- do-notation binding with a refutable pattern

All of those mechanisms resulted in an error message which identified
the line number in *Corrode's* source code where the error was detected.
None of them provided any information about the C source which triggered
the error, making it difficult to figure out what you could do about the
problem.

This commit unifies error reporting around two functions, `badSource`
and `unimplemented`, which are respectively for reporting input which we
believe to be incorrect or input which may be correct but doesn't yet
have a translation implemented.

Both error reporting functions output the C file and line number of the
problematic construct, some human-readable message, and a pretty-printed
copy of the C source in question. This is much friendlier. :-)