Output to file and parse GCC-style options.
A major change here is that the generated Rust source now goes to a file
instead of to stdout. The file has the same name as the input .c file
but with a .rs extension instead.

Together with accepting arbitrary GCC options, this allows you to run
Corrode over a whole project, typically by setting `CC=corrode`.