Use ExceptT for consistent error handling.
Now I believe nothing in this package can throw one of Haskell's
asynchronous exceptions, such as by calling `error` or `fail` or failing
a pattern-match, any of which would terminate the process.

And that changed enough of Main.hs that I decided to make it Literate
Haskell too.

I also snuck in a blank line between each generated item while I was at
it.