Reduce smaller test cases.
Requiring the short-output checksum to not be changed by test-case
reduction is too strict. It forces most global variables to be
preserved, even when they aren't contributing to differences between the
C and Rust variants of the generated program.

So now we report long-output differences only... except that even if the
long output is identical, differences in the short output also indicate
a bug. In that case we report the short-output differences.