Fix the usual arithmetic conversions.
I captured the wrong part of the pattern in `usual` when I introduced
view patterns in commit fefe39a94059a9bcc69a48811cb3d7b1ca15a98c.

The effect of this bug was that in code like `a + b`, a and b weren't
cast to their integer-promoted types if they had either the exact same
type or different signedness.

Found with csmith and creduce.