Treat va_list as a pointer to an incomplete type.
This fixes #24 about as well as I can, given that Rust doesn't support
defining variadic functions.

On every architecture I checked, __builtin_va_list is passed around as a
pointer, so this should be ABI compatible. But I haven't found anything
in the C specification requiring that, so it might be broken on some
platforms.