Look up type again if accessing incomplete structs.
If a struct was forward-declared, then completed later, we previously
didn't always allow accessing fields on that struct. It depended on
whether typedefs were involved and what order the declarations appeared
in. This commit ensures that we always find the right type declaration,
if it's in scope at the time of use.