Fix error message for typedefs of function types.
This case is not yet implemented, due to mismatches between C types,
Rust types, and the internal representation of both that I picked--but
as reported in issue #53, the resulting error message was wrong and
confusing.

This patch does not fix the underlying issue, but at least shows the
right error message in the right place.