Treat `long double` like `double`.
Rust floating-point types top out at `f64`, but the C standard has `long
double` variants of lots of functions. For the moment, just translate
this type incorrectly so we don't just die on the standard headers.

This is how Corrode has worked for a while, but I broke it in commit
92948d714b4bcf09360e40c6187eedea17f3aacd.