Strip -O flags so we translate un-optimized source.
When the build system passes GCC any of the -O family of optimization
flags, GCC sets preprocessor macros that some system headers check to
change the code they generate.

Those changes usually use GCC features that Corrode doesn't support,
such as `extern inline` (so this largely fixes #51).

But it isn't desirable to translate the "optimized" version of the
source, anyway. The unoptimized version generally expresses the
programmer's intent more clearly, so that's the version we should
translate to Rust. Optimizations after that should be the domain of the
Rust compiler, not Corrode.