Defer extern output until first use.
But once an extern is used, immediately mark its types used as well.

This commit removes the sanity-check that if an extern is re-declared,
both declarations have the same type. If you get that wrong, (1) the
generated Rust may not compile or may not work, but (2) a real C
compiler would have rejected your program anyway, so we make no promises
that Corrode will do anything sensible with it.