Emit each extern symbol at most once.
This is somewhat more efficient since it avoids an extra list traversal
per symbol lookup, and some other redundant work per use of an extern
symbol...

But more importantly, it'll be easier to ensure that anonymous types are
translated consistently under the lazy translation plan of #71 if we can
be sure they're translated only once and memoized.

This also establishes the infrastructure for carefully managed mutable
state. Rust enthusiasts should appreciate that.