Return a Result from getSymbolIdent, not a name.
Originally, I assumed that getSymbolIdent would only be used for symbols
that we translate verbatim, so an identifier in C would correspond to an
identifier in Rust. This was... naïve.

- Names declared in a C enum need to be qualified with the enum type
  name in Rust, so they translate to a 2-element path instead of a
  simple identifier.

- C builtin symbols translate to more or less complex expressions, to
  access equivalent builtin definitions from Rust's standard library.

Also, for reasons I can't remember, I stuck the Maybe in a weird place
in the return type, so that the name would be returned even if the
symbol wasn't found. However, the name is not actually used if the
symbol is not found.

Once I broadened the function's return type to include an expression,
mutability, and a type, I realized it was exactly a Result, which made
the caller in interpretExpr a lot simpler.

I also like that this makes the builtin idents more readable, since the
fields of Result are labeled.