Don't qualify enum-typed variables.
In Rust, references to enum constructors need to be qualified with the
name of the enum:

    enum E { A, B }
    E::A

But it's not enough to use the type of a symbol to decide whether its
name should be qualified this way, because that doesn't distinguish
between the constructors (like "A" and "B" above) and variables declared
of the enumerated type.

So now we record with the symbol-table entry exactly what Rust
expression will access that symbol.