Rewrite CFG into depth-first order first.
Several algorithms are simpler if we can walk the CFG in "depth-first
order", which is "the reverse of the order in which we last visit the
nodes in the preorder traversal" ("Compilers: Principles, Techniques,
and Tools", Aho/Sethi/Ullman, page 661).

Use a phantom type to ensure at compile time that we can't get confused
about whether we've already applied the depth-first ordering.

This also has the side effect of deleting blocks that are unreachable
from the function entry.