Compute loop-exit edges in depth-first order.
Identifying loop-exit edges requires that decisions for outer loops
override the decisions made for nested inner loops, while disjoint loops
have disjoint exits. But if loop A encloses loop B, the loop header for
A must come before the header for B in a depth-first ordering. So
visiting loops in depth-first order is good enough; we don't have to
work out how the loops are nested first.

This simplifies the implementation of `exitEdges` and also means
`unifyBreaks` doesn't need access to the nested hierarchy of loops.