Identify where to place each block in one pass.
This completely replaces my previous algorithm for structuring the
control graph, and has the advantage of not regressing compared to the
non-CFG version of Corrode on any source files in the CVS codebase.

In fact, as far as I can figure, this algorithm produces correct output
for every control-flow pattern, except the ones that require introducing
variables (for irreducible control flow), or introducing
single-iteration "loops" (for forward-branching early exits). Those are
both solvable problems, eventually, and in the meantime Corrode does
detect them and refuse to emit incorrect Rust.

Among many other improvements, this version fixes the earlier problem
with empty infinite loops causing Corrode to infinite-loop.

However, this version still places variable declarations in the wrong
scope, sometimes.