Partial implementation of arrays (issue #52).
Due to my kludge in commit 1118079c872eadc7336b996ec8291dd833cd2fba,
Corrode translated all array types as simple pointers. That was correct
for function arguments but utterly wrong just about everywhere else.

Because Corrode made this substitution silently, it was difficult to
spot places where the translation was wrong. So the most important thing
this commit does is ensure that Corrode will report errors when it
doesn't know how to handle array-related constructs.

In an effort to avoid regressing code that Corrode honestly was
translating successfully, this commit also explicitly transforms array
types in function arguments to the corresponding pointer types.

Finally, I took a first cut at the most common non-argument uses of
arrays, including declarations, initialization, the array-index
operator, pointer arithmetic, etc. There may well be bugs there; all I
know so far is that the output looks pretty reasonable when run on the
source code of CVS.

A major limitation of this implementation is that the array size must be
specified and must be an integer literal.