Anchor Rust `std` references to crate namespace.
Corrode emits one Rust module per C source file, but the generated
modules don't have to be used as a stand-alone Rust crate. If one of
these modules is nested inside a larger crate, then references to the
Rust `std` crate need to be relative to the crate root, not to the
current module.