CFG: preserve entry sets at each block.
By keeping track of whether a branch refers to a label in the entry set
of the lexically following block, we can avoid inserting `continue`
statements that branch to where control would have gone anyway.

Furthermore, I believe the exact requirement for eliding a `mkGoto`
assignment is that the block we're jumping to has exactly one entry
label. So this change allows expressing that rule precisely.

This also enables a runtime check verifying that falling off the end of
a block will continue to the right next block. Without knowing the entry
labels that the next block was constructed for, we can't tell by looking
at it whether it's prepared to handle the label we want to branch to.