Labels not actually in the handler are also out-edges.
When we're constructing a Multiple block, some of the blocks inside it
may have branches to labels that are outside the Multiple block. Those
may naively appear to be singly-reached labels, but since those external
blocks can't be placed in the corresponding branch of the Multiple
block, we need to treat them as out-edges instead.

Fixes test reproducible with:

--quickcheck-replay '35 TFGenR 564FE93137373F75BE33AC8B09A45168950125A98019EBFE8EC2986A203EB500 0 255 8 0'