Document Relooper cases and termination argument.
In the process I realized that the Emscripten paper describes
correctness conditions that are more complicated than necessary,
allowing me to simplify this code a little bit. Because no-return entry
labels are a subset of singly-reachable entry labels, if we know there's
at least one no-return entry label then we know we can safely construct
at least one handler in a `Multiple` block.