Fix "if may be missing an else clause" error.
In some cases, the if-else ladders generated by structuring the
control-flow graph are guaranteed to have some branch evaluate true.
However, the Rust compiler couldn't prove that. As a result the
compiler's reachability analysis was overly conservative, leading to
error messages about paths that we know couldn't happen.

Fixed by replacing the final `if` condition in such blocks with an
unconditional `else`.

Fixes #134.