AST: Fix parenthesis for blocks in if-statements.
In Rust syntax, any block that begins at the start of a statement is
parsed as a statement, even if it is not followed by a semicolon. If you
want such a block to be parsed as an expression instead, you need to
wrap parentheses around it. The expression at the end of the body of an
`if`-expression is, naturally, a case that should be parsed as an
expression, so ensure we insert parentheses appropriately.